% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headless-package.R
\docType{package}
\name{headless-package}
\alias{headless}
\alias{headless-package}
\title{headless: Headless integration tests for extendR}
\description{
Suited for internal tests of extendr, where we may import `libR-sys`
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
